requirements.txt
streamlit==1.37.1
langchain==0.2.12
langchain-core==0.2.36
langchain-community==0.2.10
langchain-openai==0.1.22
openai>=1.30.0,<2
tiktoken==0.7.0
faiss-cpu==1.8.0.post1
pymupdf==1.24.8
python-docx==1.1.2
docx2txt==0.8
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1